Add support for music and sounds
* Add SoundScheme helper for playing a sound


* Add support for changing sound scheme via settings


* Add support for sound schemes


* Update .gitignore


* Fix sliders only generatibg change events on whole numbers


* Properly apply music volume to new tracks as they start playing


* Add support for setting music volume


* Add support for music playback via PlaySongEvent

+++ src/AcidicGUI/Widgets/Slider.cs
@@ -90,11 +90,11 @@ public sealed class Slider : Widget,
 
         if (direction == Direction.Vertical)
         {
-            SetValue(1 - MathHelper.Clamp(e.Position.Y - ContentArea.Top, 0, ContentArea.Height) / ContentArea.Height);
+            SetValue(1 - MathHelper.Clamp(e.Position.Y - ContentArea.Top, 0, ContentArea.Height) / (float) ContentArea.Height);
         }
         else
         {
-            SetValue(MathHelper.Clamp(e.Position.X - ContentArea.Left, 0, ContentArea.Width) / ContentArea.Width);
+            SetValue(MathHelper.Clamp(e.Position.X - ContentArea.Left, 0, ContentArea.Width) / (float) ContentArea.Width);
         }
         
         InvalidateGeometry();
