+++ src/SociallyDistant.Framework/Core/Config/ISettingsUiBuilder.cs
@@ -9,7 +9,7 @@ namespace SociallyDistant.Core.Core.Config
 		ISettingsUiBuilder WithLabel(string labelText, int sectionId);
 		ISettingsUiBuilder WithToggle(string title, string? description, bool value, Action<bool> changeCallback, int sectionId);
 		ISettingsUiBuilder WithSlider(string title, string? description, float value, float minimum, float maximum, Action<float> changeCallback, int sectionId);
-		ISettingsUiBuilder WithSlider(string title, string? description, float value, int minimum, int maximum, Action<int> changeCallback, int sectionId);
+		ISettingsUiBuilder WithSlider(string title, string? description, int value, int minimum, int maximum, Action<int> changeCallback, int sectionId);
 		ISettingsUiBuilder WithTextField(string title, string? description, string? currentValue, Action<string?> changeCallback, int sectionId);
 		ISettingsUiBuilder WithStringDropdown(string title, string? description, int currentIndex, string[] choices, Action<int> changeCallback, int sectionId);
 	}
